Add HttpContextAccessor

+++ backend/net-core-mssql/Services/CharacterService.cs
@@ -1,8 +1,10 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using net_core_mssql.Data;
 using net_core_mssql.Dtos.Character;
@@ -14,9 +16,12 @@ namespace net_core_mssql.Services
   {
     private readonly IMapper mapper;
     private readonly DataContext context;
+    private readonly IHttpContextAccessor httpContextAccessor;
+    private int GetUserId() => int.Parse(this.httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
-    public CharacterService(IMapper mapper, DataContext context)
+    public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
     {
+      this.httpContextAccessor = httpContextAccessor;
       this.context = context;
       this.mapper = mapper;
     }
@@ -25,10 +30,12 @@ namespace net_core_mssql.Services
     {
       ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
       Character character = mapper.Map<Character>(newCharacter);
+      character.User = await context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
 
       await context.Characters.AddAsync(character);
       await context.SaveChangesAsync();
-      serviceResponse.Data = (context.Characters.Select(c => mapper.Map<GetCharacterDto>(c))).ToList();
+
+      serviceResponse.Data = (context.Characters.Where(c => c.User.Id == GetUserId()).Select(c => mapper.Map<GetCharacterDto>(c))).ToList();
       return serviceResponse;
     }
 
@@ -37,10 +44,20 @@ namespace net_core_mssql.Services
       ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
       try
       {
-        Character character = await context.Characters.FirstAsync(c => c.Id == id);
-        context.Characters.Remove(character);
-        await context.SaveChangesAsync();
-        serviceResponse.Data = (context.Characters.Select(c => mapper.Map<GetCharacterDto>(c))).ToList();
+        Character character = 
+            await context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+        if (character != null)
+        {
+            context.Characters.Remove(character);
+            await context.SaveChangesAsync();
+            serviceResponse.Data = (context.Characters.Where(c => c.User.Id == GetUserId())
+                .Select(c => mapper.Map<GetCharacterDto>(c))).ToList();
+        }
+        else
+        {
+            serviceResponse.Success = false;
+            serviceResponse.Message = "Character not found.";
+        }
       }
       catch (Exception ex)
       {
@@ -50,10 +67,10 @@ namespace net_core_mssql.Services
       return serviceResponse;
     }
 
-    public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+    public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
     {
       ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-      List<Character> dbCharacters = await context.Characters.Where(c => c.User.Id == userId).ToListAsync();
+      List<Character> dbCharacters = await context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
       serviceResponse.Data = dbCharacters.Select(c => mapper.Map<GetCharacterDto>(c)).ToList();
       return serviceResponse;
     }
@@ -61,7 +78,7 @@ namespace net_core_mssql.Services
     public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
     {
       ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
-      Character dbCharacter = await context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+      Character dbCharacter = await context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
       serviceResponse.Data = mapper.Map<GetCharacterDto>(dbCharacter);
       return serviceResponse;
     }
@@ -71,18 +88,27 @@ namespace net_core_mssql.Services
       ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
       try
       {
-        Character character = await context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
-        character.Name = updatedCharacter.Name;
-        character.Class = updatedCharacter.Class;
-        character.Defense = updatedCharacter.Defense;
-        character.HitPoints = updatedCharacter.HitPoints;
-        character.Intelligence = updatedCharacter.Intelligence;
-        character.Strength = updatedCharacter.Strength;
+        // So keep that in mind, if you want to access related objects. You might have to include them first.
+        Character character = await context.Characters.Include(c => c.User).FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+        if (character != null && character.User.Id == GetUserId())
+        {
+            character.Name = updatedCharacter.Name;
+            character.Class = updatedCharacter.Class;
+            character.Defense = updatedCharacter.Defense;
+            character.HitPoints = updatedCharacter.HitPoints;
+            character.Intelligence = updatedCharacter.Intelligence;
+            character.Strength = updatedCharacter.Strength;
 
-        context.Characters.Update(character);
-        await context.SaveChangesAsync();
+            context.Characters.Update(character);
+            await context.SaveChangesAsync();
 
-        serviceResponse.Data = mapper.Map<GetCharacterDto>(character);
+            serviceResponse.Data = mapper.Map<GetCharacterDto>(character);
+        }
+        else
+        {
+            serviceResponse.Success = false;
+            serviceResponse.Message = "Character not found.";
+        }
       }
       catch (Exception ex)
       {
